Refactor for compatibility with Jenkins Pipeline jobs
This commit refactors the plugin to use more modern Jenkins APIs,
so that it is compatible with the new "Pipeline" jobs.

This means that you can include a line like this in your Jenkins Pipeline
script:

    gatlingArchive()

And the effect will be the same as if you'd added the Gatling post-build
action to your freestyle job; that is, you'll see the Gatling link appear
in the sidebar of both the Project and Build pages, the reports will be
available, the graphs will be visible, etc.